+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ## [Unreleased]
+### Fixed
+* Copying images to clipboard would fail with an xclip error
 
 
 ## [1.4.0] - 2020-07-09